% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{Lall}
\alias{Lall}
\title{Hypothesis matrix to test for lmer objects}
\usage{
Lall(fit, nam)
}
\arguments{
\item{fit}{a fitted model with a 'getFix' method.}
}
\value{

}
\description{
Hypothesis matrix to test for lmer objects
}
