% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{Ldiff.old}
\alias{Ldiff.old}
\title{Older version of Ldiff}
\usage{
Ldiff.old(fit, pat, levnames = c(reflevel, substring(rownames(L), cut +
  1)), reflevel = "<ref>", cut = nchar(pat))
}
\arguments{
\item{fit}{a fitted model with a 'getFix' method.}

\item{pat}{a regular expression that matches names of coefficients}

\item{levnames}{level names; levnames = c(reflevel,substring(rownames(L),cut+1)}

\item{reflevel}{reference level to test against}
}
\value{
hypothesis matrix
}
\description{
Older version of Ldiff
}
