% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{Ldiff.rdc}
\alias{Ldiff.rdc}
\title{Version of Ldiff used in RDC}
\usage{
Ldiff.rdc(fit, nam, ref = "no longer used")
}
\arguments{
\item{fit}{a fitted model with a 'getFix' method.}

\item{ref}{reference level to test against}
}
\value{
hypothesis matrix
}
\description{
Version of Ldiff used in RDC
}
