% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{Lequal}
\alias{Lequal}
\title{Hypothesis matrix to test equality of factor level effects}
\usage{
Lequal(fit, pat)
}
\arguments{
\item{fit}{a fitted model with a 'getFix' method.}

\item{pat}{a regular expression that matches names of coefficients}
}
\value{
hypothesis matrix
}
\description{
Hypothesis matrix to test equality of factor level effects
}
