% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{Lmu}
\alias{Lmu}
\title{Estimate predicted response for a factor level.}
\usage{
Lmu(fit, nam, verbose = 0)
}
\arguments{
\item{fit}{a fitted model with a 'getFix' method.}
}
\description{
Estimate predicted response for a factor level.
}
