% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{print.wald}
\alias{print.wald}
\title{Print method for wald objects}
\usage{
\method{print}{wald}(x, round = 6, pround = 5, ...)
}
\arguments{
\item{x}{(TO DO: GEORGES)}

\item{round}{(TO DO: GEORGES)}

\item{pround}{(TO DO: GEORGES)}

\item{\dots}{(TO DO: GEORGES)}
}
\value{
called for
}
\description{
Print method for wald objects
}
\note{
This is a note
}
\examples{
# coming soon
}
\seealso{
\code{\link{wald}}
}
\author{
GM
}
