% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tran.R
\name{tr}
\alias{tr}
\title{Translate elements of a character vector}
\usage{
tr(x, from, to)
}
\arguments{
\item{x}{vector to apply translation to}

\item{from}{vector to translate from}

\item{to}{vector to translate to}
}
\description{
Changes the order of arguments in \code{\link{tran}}.
}
